fix incognito file not found bug
Previous builds used window.open, which would open the screenshot in a new non-incognito window, so this just reverts that same type of functionality, but using the more powerful chrome.tabs and chrome.windows APIs.

Also, put in a smarter tab indexing, so new tabs show up next to the current tab (when not in incognito), instead of at the end.